+++ lib/ui/blocs/posts/export.dart
@@ -1,4 +1,3 @@
 export 'post_details/bloc.dart';
-export 'post_input/bloc.dart';
 export 'posts_list/bloc.dart';
 export 'report_popup/bloc.dart';
