+++ lib/ui/screens/post_create_screen/blocs/post_input/post_input_event_poll.dart
@@ -1,4 +1,4 @@
-import 'package:mooncake/ui/ui.dart';
+import './export.dart';
 
 /// Represents a generic event related to the poll of a post.
 abstract class PostInputPollEvent extends PostInputEvent {}
